moved editor switching tests within supereditor_multi_editor_test
